+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [unreleased]
 ### Added
+- Added UnderstandImporter to Analysis
 
 ### Changed
 
@@ -35,6 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Fixed
 - filter by regex shows parent nodes #116
+- typo in scss file
 
 ## [1.12.0] - 2018-04-27
 ### Added